# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NotaFácil API",
    "version": "1.0.0",
    "description": "API do NotaFácil: registro de notas fiscais eletrônicas de consumo e comparação do histórico de preços do próprio usuário.\n\nO registro ocorre por três caminhos: leitura do QR Code, digitação da chave de acesso e importação do XML autorizado. Quando a consulta automática não está disponível, a API responde 422 com o código `CONSULTA_ASSISTIDA` e o endereço da consulta oficial da SEFAZ.\n\nTodas as consultas são restritas ao usuário autenticado: notas, produtos, preços por estabelecimento e comparativos nunca incluem compras de outros usuários. As rotas de conta implementam os direitos do titular previstos na LGPD (art. 18).",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/api/auth/cadastro": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Cria uma conta",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nome",
                  "email",
                  "senha",
                  "aceitePolitica",
                  "consentimentoDadosSensiveis",
                  "versaoPolitica"
                ],
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "senha": {
                    "type": "string",
                    "minLength": 8
                  },
                  "aceitePolitica": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "consentimentoDadosSensiveis": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  },
                  "versaoPolitica": {
                    "type": "string",
                    "description": "Versão vigente, obtida em /api/privacidade."
                  }
                }
              },
              "example": {
                "nome": "Maria Souza",
                "email": "maria@exemplo.com",
                "senha": "segredo123",
                "aceitePolitica": true,
                "consentimentoDadosSensiveis": true,
                "versaoPolitica": "1.0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Conta criada, já autenticada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Autenticacao"
                }
              }
            }
          },
          "403": {
            "description": "Cadastro fechado nesta instalação (CADASTRO_ABERTO=false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "erro": "O cadastro de novas contas está fechado nesta instalação."
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflito"
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          },
          "429": {
            "description": "Limite de contas criadas pelo mesmo IP na última hora (LIMITE_CADASTROS_POR_HORA, padrão 5)",
            "headers": {
              "Retry-After": {
                "description": "Segundos até liberar novos cadastros",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "erro": "Muitas contas criadas a partir desta rede. Tente novamente em 42 minuto(s)."
                }
              }
            }
          }
        },
        "description": "Exige o aceite da política de privacidade vigente e o consentimento específico para compras que possam revelar dados de saúde (art. 11, I, da LGPD). A versão aceita e a data do aceite ficam registradas na conta."
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Entra no sistema",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "senha"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "senha": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "email": "demo@notafacil.app",
                "senha": "demo1234"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token de sessão, válido por 24 horas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Autenticacao"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          },
          "429": {
            "description": "Muitas tentativas de login",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos até liberar novas tentativas."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "description": "Depois de 5 tentativas com senha errada para o mesmo e-mail e origem em 15 minutos, novas tentativas são recusadas com 429 até a janela expirar."
      }
    },
    "/api/conta": {
      "get": {
        "tags": [
          "Conta e privacidade"
        ],
        "summary": "Dados da conta e situação do aceite da política",
        "responses": {
          "200": {
            "description": "Conta do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conta"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          }
        }
      },
      "patch": {
        "tags": [
          "Conta e privacidade"
        ],
        "summary": "Corrige nome ou e-mail (direito de correção)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conta atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conta"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "409": {
            "$ref": "#/components/responses/Conflito"
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          }
        }
      },
      "delete": {
        "tags": [
          "Conta e privacidade"
        ],
        "summary": "Exclui a conta e todos os dados (direito de eliminação)",
        "description": "Apaga a conta, as notas, os itens e os preços do usuário. Produtos e estabelecimentos que não constem em notas de outros usuários também são removidos. O token deixa de valer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "senha"
                ],
                "properties": {
                  "senha": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Conta excluída"
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "403": {
            "description": "Senha incorreta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          }
        }
      }
    },
    "/api/conta/exportacao": {
      "get": {
        "tags": [
          "Conta e privacidade"
        ],
        "summary": "Exporta todos os dados do titular em JSON (acesso e portabilidade)",
        "responses": {
          "200": {
            "description": "Arquivo JSON com conta, notas, itens e preços por estabelecimento, sem a senha",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exportacao"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          }
        }
      }
    },
    "/api/conta/politica": {
      "post": {
        "tags": [
          "Conta e privacidade"
        ],
        "summary": "Aceita a versão vigente da política",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "versao",
                  "consentimentoDadosSensiveis"
                ],
                "properties": {
                  "versao": {
                    "type": "string"
                  },
                  "consentimentoDadosSensiveis": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Aceite registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conta"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          }
        }
      }
    },
    "/api/notas/qrcode": {
      "post": {
        "tags": [
          "Notas"
        ],
        "summary": "Registra uma nota pelo QR Code ou pela chave de acesso",
        "description": "Aceita os 44 dígitos da chave, a chave com separadores ou a URL completa gravada no QR Code. A chave é validada pelo módulo 11 antes de qualquer consulta, e a duplicidade é verificada antes da consulta paga.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "conteudo"
                ],
                "properties": {
                  "conteudo": {
                    "type": "string",
                    "description": "Chave de acesso ou conteúdo lido do QR Code."
                  }
                }
              },
              "example": {
                "conteudo": "25260307526557000100650010000010011101371370"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nota registrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notaId": {
                      "type": "integer"
                    },
                    "chaveInterpretada": {
                      "$ref": "#/components/schemas/ChaveInterpretada"
                    },
                    "nota": {
                      "$ref": "#/components/schemas/NotaDetalhe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "409": {
            "$ref": "#/components/responses/Conflito"
          },
          "422": {
            "description": "Chave inválida, nota cancelada ou consulta automática indisponível",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "examples": {
                  "chaveInvalida": {
                    "summary": "Dígito verificador não confere",
                    "value": {
                      "erro": "Chave de acesso inválida: o dígito verificador não confere."
                    }
                  },
                  "consultaAssistida": {
                    "summary": "Portal exige verificação humana",
                    "value": {
                      "erro": "A consulta automática não retornou esta nota. Abra a consulta oficial da SEFAZ-SP, que pode exigir verificação humana, e importe o XML autorizado.",
                      "codigo": "CONSULTA_ASSISTIDA",
                      "urlConsulta": "https://www.sefaz.pb.gov.br/nfce?p=25260307526557000100650010000010011101371370"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notas/xml": {
      "post": {
        "tags": [
          "Notas"
        ],
        "summary": "Registra uma nota a partir do XML autorizado",
        "description": "O XML pode ser enviado como corpo `application/xml` ou dentro do campo `xml` de um JSON.",
        "requestBody": {
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "xml": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nota registrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notaId": {
                      "type": "integer"
                    },
                    "nota": {
                      "$ref": "#/components/schemas/NotaDetalhe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "409": {
            "$ref": "#/components/responses/Conflito"
          },
          "422": {
            "$ref": "#/components/responses/DadosInvalidos"
          }
        }
      }
    },
    "/api/notas": {
      "get": {
        "tags": [
          "Notas"
        ],
        "summary": "Lista as notas do usuário",
        "parameters": [
          {
            "name": "chave",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Trecho da chave de acesso."
          },
          {
            "name": "dataInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dataFim",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "empresaId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "busca",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Razão social, nome fantasia ou CNPJ."
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade e totais de todas as notas filtradas, com a lista limitada às 200 mais recentes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListagemNotas"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          }
        }
      }
    },
...
    "/api/notas/{id}/comparacao": {
      "get": {
        "tags": [
          "Notas"
        ],
        "summary": "Compara os itens da nota com o histórico",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Itens com situação, preço anterior, menor preço e economia possível",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemComparado"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NaoAutorizado"
          },
          "404": {
            "$ref": "#/components/responses/NaoEncontrado"
          }
        },
        "description": "Cada item é comparado apenas com compras do mesmo produto feitas antes da nota, pela data e hora de emissão."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token devolvido por /api/auth/login ou /api/auth/cadastro."
      }
    },
...
    "schemas": {
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string"
          },
          "codigo": {
            "type": "string",
            "description": "Presente em falhas previstas. `CONSULTA_ASSISTIDA` indica que o usuário deve consultar o portal oficial e importar o XML."
          },
          "urlConsulta": {
            "type": "string",
            "nullable": true,
            "description": "Endereço da consulta oficial da SEFAZ, quando há um para a unidade federativa da nota."
          }
        },
        "required": [
          "erro"
        ]
      },
...
      "Autenticacao": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT válido por sete dias."
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}